Add Reminder Function
Date and time entered by the user for deadline and event tasks are not in use properly.

Create reminder functions that allows user to key in dates and find tasks that will be due by that date.

Helps to make a better use of the date and time entered by the user, making more sense of the user input